1.3, mostly qol and bug fixes but also adds late joiners to the start message

+++ scripts/gameReset.js
@@ -1,4 +1,4 @@
-const { staffChannelId, picksChannelId, guildId } = require('./../config.json');
+const { staffChannelId, picksChannelId, guildId, captainId } = require('../config.json');
 const schedule = require('node-schedule');
 const { QuickDB } = require("quick.db");
 const db = new QuickDB();
@@ -7,19 +7,22 @@ const { AttachmentBuilder } = require("discord.js");
 
 module.exports = {
     async execute(client){
-        const rule = new schedule.RecurrenceRule(); // 7pm EDT on Saturday (11pm UTC)
-        rule.dayOfWeek = 0;
-        rule.hour = 0;
-        rule.minute = 0;
-        rule.second = 0;
+        const rule = new schedule.RecurrenceRule(); // 12am EDT on Sunday (4am UTC)
+        rule.dayOfWeek = await db.get(`startDay`);
+        rule.hour = 23;
+        rule.minute = 59;
+        rule.second = 59;
         rule.tz = 'Etc/GMT+4';
 
         const job = schedule.scheduleJob(rule, async function(){
+            if(await db.get(`started`) == false) return;
+
             // saves and deletes all messages in #picks
             let output = "";
             await client.channels.cache.get(picksChannelId).messages.fetch({ limit: 100 }).then(messages => {
-                messages.forEach(msg => {
-                    output += `${msg.author.username}: ${msg.content}\n\n`
+                console.log(messages)
+                messages.reverse().forEach(msg => {
+                    output += `${msg.author.username}: ${msg.content}\n\n`;
                 });
 
                 client.channels.cache.get(picksChannelId).bulkDelete(messages).catch(() => {
@@ -32,14 +35,17 @@ module.exports = {
             // reassigns captains their original names
             try{
                 const nicknames = await db.get('captains');
+                const captain = await interaction.guild.roles.cache.get(captainId);
 
                 await client.guilds.cache.get(guildId).members.fetch({ limit: 1000 }).then(members => {
                     for(let obj of nicknames){
                         members.get(obj.id).setNickname(obj.nickname);
+                        members.get(obj.id).roles.remove(captain);
                     }
                 });
     
                 await db.set('captains', []);
+                await db.set('started', false);
             }catch(err){
                 console.error(err);
             }
@@ -48,15 +54,15 @@ module.exports = {
             const week = await db.get(`currentWeek`);
 
             // writes the file to the directory
-            fs.writeFileSync(`./out/picks/s${season}w${week}.txt`, Buffer.from(output), function(){});
+            fs.writeFileSync(`./out/picks/s${season}w${week}_picks.txt`, Buffer.from(output), function(){});
 
-            const attachment = new AttachmentBuilder(`out/picks/s${season}w${week}.txt`);
+            const attachment = new AttachmentBuilder(`out/picks/s${season}w${week}_picks.txt`);
 
             client.channels.cache.get(staffChannelId).send({ content: `Picks from season \`${season}\` week \`${week}\`:`, files: [ attachment ] });
 
-            console.log(`Next Queue: ${job.nextInvocation().toISOString()}`);
+            console.log(`Next Reset: ${job.nextInvocation().toISOString()}`);
         });
 
-        console.log(`Next Queue: ${job.nextInvocation().toISOString()}`);
+        console.log(`Next Reset: ${job.nextInvocation().toISOString()}`);
     }
 }
